# ENH: Series.resample performance with datetime64[ns] #7754