--- a/doc/source/whatsnew/v0.17.0.txt
+++ b/doc/source/whatsnew/v0.17.0.txt
@@ -53,6 +53,8 @@ Removal of prior version deprecations/changes
 Performance Improvements
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
+- Improved ``Series.resample`` performance with dtype=datetime64[ns] (:issue:`7754`)
+
 .. _whatsnew_0170.bug_fixes:
 
 Bug Fixes