It fixes the performance drop.
See %timeit result [here](http://nbviewer.ipython.org/gist/scari/1173a9519d8b89000c08).

@jreback Would you review this PR? Thanks!