@jreback 2015-05-04
hmm, would need a benchmark, e.g. post a sample timeit for before/after (on a good sized test frame)