@scari 2015-05-04
```python
import pandas as pd

intSeries = pd.Series(5, pd.date_range(start='2000-01-01', end='2000-01-08', freq='555000U'), dtype='int64')
timeSeries = intSeries.astype('datetime64[ns]')

%timeit intSeries.resample('1S', how='last')
%timeit timeSeries.resample('1S', how='last')
%prun intSeries.resample('1S', how='last')
%prun timeSeries.resample('1S', how='last')
```

### Before ###
100 loops, best of 3: 18.8 ms per loop
1 loops, best of 3: **17 s** per loop
449 function calls (441 primitive calls) in 0.021 seconds
[snip]
22378318 function calls (21773499 primitive calls) in 24.219 seconds
[snip]

### After ###
100 loops, best of 3: 19 ms per loop
100 loops, best of 3: **20 ms** per loop
449 function calls (441 primitive calls) in 0.020 seconds
[snip]
418 function calls (410 primitive calls) in 0.020 seconds
[snip]