@jreback 2015-05-04
ok, looks ok, can you add a vbench in ``vb_suite/timeseries.py``. you can make similar to the above. but make smaller, so that the slowest time doesn't take too long (max 1s)