@scari 2015-05-05
Added test. @jreback please review again.
```
$ ./test_perf.sh -b HEAD~1 -t HEAD -r timeseries
-------------------------------------------------------------------------------
Test name                                    | head[ms] | base[ms] |  ratio   |
-------------------------------------------------------------------------------
timeseries_resample_datetime64               |   3.7426 | 2009.6250 |   0.0019 |

Target [dd6ffa3] : ENH: Series.resample performance with datetime64[ns] #7754
Base   [3128215] : cleanup test for GH10044
```