@scari 2015-05-12
Updated whatsnew/v0.17.0.txt.